Use qualname to retrieve matching routing entries.
Functions as keys are not found because each Rafter filter will alter
the function id, so we use the qualname to found the function which
match the same qualname.

This will not work if 2 functions have the same name in the codebase
but it's the best we can do for now.